api/app.py
from flask import Flask, jsonify
import psycopg2
import yaml
import os

app = Flask(__name__)

# Load configuration from YAML
# Implement data transformation logic here based on how you provide ConfigMap to service

# ...

# Database connection details
DATABASE_URL = os.getenv('DATABASE_URL', 'postgres://postgres:password@postgres-service:5432/mydb')

@app.route("/<string:path>")
@app.route("/<path:path>")

def index(path):
  try:
    conn = psycopg2.connect(DATABASE_URL)
    cur = conn.cursor()
    # Generate SQL query based on configured table name for the path
    # Implement here
    # ...

    # query = cur.execute(query)
    data = cur.fetchall()
    cur.close()
    conn.close()

    return jsonify(transform_data(data))
  except Exception as e:
    return jsonify({'error': str(e)}), 500

@app.route("/health")
def health():
    try:
        conn = psycopg2.connect(DATABASE_URL)
        cur = conn.cursor()
        cur.execute('SELECT 1')
        cur.close()
        conn.close()
        return jsonify({'status': 'healthy'}), 200
    except Exception as e:
        return jsonify({'status': 'unhealthy', 'error': str(e)}), 500

def transform_data(data):
  # Implement data transformation logic here.
  # i.e. filter out undesired columns, remap column names to object fields
  # ...

  return data

if __name__== '__main__':
  app.run(debug=True, host='0.0.0.0', port=3000)